+++ src/main/java/svenhjol/charmony/core/common/mixins/anvil/AnvilMenuRepairMixin.java
@@ -1,4 +1,4 @@
-package svenhjol.charmony.core.common.mixins.anvil_repair;
+package svenhjol.charmony.core.common.mixins.anvil;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.sugar.Local;
@@ -8,18 +8,18 @@ import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import svenhjol.charmony.core.events.AnvilRepairEvent;
+import svenhjol.charmony.core.events.AnvilEvents;
 
 @SuppressWarnings("UnreachableCode")
 @Mixin(AnvilMenu.class)
-public abstract class AnvilMenuMixin extends ItemCombinerMenu {
-    public AnvilMenuMixin(@Nullable MenuType<?> menuType, int i, Inventory inventory, ContainerLevelAccess access,
-                          ItemCombinerMenuSlotDefinition definition) {
+public abstract class AnvilMenuRepairMixin extends ItemCombinerMenu {
+    public AnvilMenuRepairMixin(@Nullable MenuType<?> menuType, int i, Inventory inventory, ContainerLevelAccess access,
+                                ItemCombinerMenuSlotDefinition definition) {
         super(menuType, i, inventory, access, definition);
     }
 
     /**
-     * Fires the {@link AnvilRepairEvent} event.
+     * Fires the {@link svenhjol.charmony.core.events.AnvilEvents.RepairEvent} event.
      * Allows intervention when checking if the anvil item can be repaired with another item via the CheckAnvilRepairEvent event.
      * For example, elytra cannot normally be repaired with leather, but using the player.world we can check if insomnia is disabled
      * and therefore allow this repair check.  The reason we don't hook into ElytraItem's canRepair method directly is because
@@ -37,7 +37,7 @@ public abstract class AnvilMenuMixin extends ItemCombinerMenu {
                                               @Local(ordinal = 1) ItemStack leftItem,
                                               @Local(ordinal = 2) ItemStack rightStack
     ) {
-        boolean result = AnvilRepairEvent.INSTANCE.invoke((AnvilMenu) (Object) this, this.player, leftStack, rightStack);
+        boolean result = AnvilEvents.REPAIR.invoke((AnvilMenu) (Object) this, this.player, leftStack, rightStack);
         return result || original;
     }
 }